Start a `util` namespace
I wanted my grammar work to be in its own
namespace, and it turned out that I had
dependencies on things in `core`. This
required moving some of the "utility"
functions into a new `util` namespace so
they can be required by both `core` and the
grammar work.

I'm not 100% sure these are really `util`
things – it might make more sense to have
a `push-state` namespace and put them
there.