Move more functions to `util`
It turns out that most of the interpreter
code needs to be in `util` so the grammar
code can access them. `util` really needs
to be split up. There are real "util"
things here like `abs`, but there's a lot
of push state stuff that should be in its
own namespace.